Added word scoring and simple matching
Simple matching is awesome and allows you to input existing letters,
spacing, and edges as they exist on the board. This match is then
translated into a regular expression for you.

[F_X] -> ^F[a-z]X$
F_X]  -> .*F[a-z]X$
F_X   -> .*F[a-z]X
[F_X  -> F[a-z]X

That is, [ and ] (and also, ^ and $, and | and |) can be used to
mark a word edge. That is, no letters can come before or after
these characters, respectively.

_ (and also ? and *) can be used to mark a wildcard character.
Any character can be used at that portion of the word.